+++ test/sql/class.sql
@@ -57,7 +57,7 @@ SELECT ddlx_script('test_class_v'::regtype);
 
 CREATE VIEW test_class_v2 AS
 SELECT * FROM test_class_v;
-grant select on test_class_v2 to public;
+grant select (a,b,c) on test_class_v2 to public;
 SELECT ddlx_script('test_class_v'::regclass);
 
 CREATE MATERIALIZED VIEW test_class_m AS