+++ CHANGELOG.md
@@ -2,8 +2,8 @@
 
 Version 0.11
 ------------
-- function ddlx_get_dependants_recursive() rolled into ddlx_get_dependants()
 - support for column grants
+- function ddlx_get_dependants_recursive() rolled into ddlx_get_dependants()
 
 Version 0.10
 ------------